Extract SubscriptionExecutor into its own file
in preparation for merging with Executor
